--- 3dreact/src/App.css
@@ -1,25 +1,0 @@
-.body {
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  margin: 0;
-  font-family: Arial, sans-serif;
-  font-size: 16px;
-  color: #333;
-  background-color: #f4f4f4;
-}
-
-
-.top {
-  background-color: #1e1e1e;
-  color: white;
-  padding: 5px;
-  text-align: center;
-}
-
-.main {
-  display: flex;
-  width: 100%;
-  flex-grow: 1;
-  position: relative;
-}
