+++ frontend/package.json
@@ -11,6 +11,8 @@
     "react-router-dom": "^6.23.1",
     "react-scripts": "5.0.1",
     "react-transition-group": "^4.4.5",
+    "three": "^0.165.0",
+    "tree": "^0.1.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
