--- frontend/src/components/canvasComponent.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function CanvasComponent(props) {
-  return (
-    <div>
-        <canvas className = "StellarSystem"></canvas>
-    </div>
-  );
-}
-
-export default CanvasComponent;
