+++ frontend/src/pages/UI.css
@@ -30,3 +30,15 @@ p {
     left: 0;
     padding: 10px;
 }
+
+.topCanvas {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    z-index: 1000;
+}
