+++ frontend/src/utils/resizeHandler.js
