+++ frontend/src/utils/sceneSetup.js
